updated Glue examples
