--- javav2/example_code/glue/src/main/java/com/example/glue/GetDatabase.java
@@ -1,77 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.example.glue;
-
-// snippet-start:[glue.java2.get_database.main]
-// snippet-start:[glue.java2.get_database.import]
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.glue.GlueClient;
-import software.amazon.awssdk.services.glue.model.GetDatabaseRequest;
-import software.amazon.awssdk.services.glue.model.GetDatabaseResponse;
-import software.amazon.awssdk.services.glue.model.GlueException;
-import java.time.Instant;
-import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
-import java.time.format.FormatStyle;
-import java.util.Locale;
-// snippet-end:[glue.java2.get_database.import]
-
-/**
- * Before running this Java V2 code example, set up your development
- * environment, including your credentials.
- *
- * For more information, see the following documentation topic:
- *
- * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
- */
-public class GetDatabase {
-    public static void main(String[] args) {
-        final String usage = """
-
-                Usage:
-                    <databaseName>
-
-                Where:
-                    databaseName - The name of the database.\s
-                """;
-
-        if (args.length != 1) {
-            System.out.println(usage);
-            System.exit(1);
-        }
-
-        String databaseName = args[0];
-        Region region = Region.US_EAST_1;
-        GlueClient glueClient = GlueClient.builder()
-                .region(region)
-                .build();
-
-        getSpecificDatabase(glueClient, databaseName);
-        glueClient.close();
-    }
-
-    public static void getSpecificDatabase(GlueClient glueClient, String databaseName) {
-        try {
-            GetDatabaseRequest databasesRequest = GetDatabaseRequest.builder()
-                    .name(databaseName)
-                    .build();
-
-            GetDatabaseResponse response = glueClient.getDatabase(databasesRequest);
-            Instant createDate = response.database().createTime();
-
-            // Convert the Instant to readable date.
-            DateTimeFormatter formatter = DateTimeFormatter.ofLocalizedDateTime(FormatStyle.SHORT)
-                    .withLocale(Locale.US)
-                    .withZone(ZoneId.systemDefault());
-
-            formatter.format(createDate);
-            System.out.println("The create date of the database is " + createDate);
-
-        } catch (GlueException e) {
-            System.err.println(e.awsErrorDetails().errorMessage());
-            System.exit(1);
-        }
-    }
-}
-// snippet-end:[glue.java2.get_database.main]
