--- javav2/example_code/glue/src/main/java/com/example/glue/GetTable.java
@@ -1,80 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.example.glue;
-
-// snippet-start:[glue.java2.get_table.main]
-// snippet-start:[glue.java2.get_table.import]
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.glue.GlueClient;
-import software.amazon.awssdk.services.glue.model.GetTableRequest;
-import software.amazon.awssdk.services.glue.model.GetTableResponse;
-import software.amazon.awssdk.services.glue.model.GlueException;
-import java.time.Instant;
-import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
-import java.time.format.FormatStyle;
-import java.util.Locale;
-// snippet-end:[glue.java2.get_table.import]
-
-/**
- * Before running this Java V2 code example, set up your development
- * environment, including your credentials.
- *
- * For more information, see the following documentation topic:
- *
- * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
- */
-public class GetTable {
-    public static void main(String[] args) {
-        final String usage = """
-
-                Usage:
-                    <dbName> <tableName>
-
-                Where:
-                    dbName - The database name.\s
-                    tableName - The name of the table.\s
-                """;
-
-        if (args.length != 2) {
-            System.out.println(usage);
-            System.exit(1);
-        }
-
-        String dbName = args[0];
-        String tableName = args[1];
-        Region region = Region.US_EAST_1;
-        GlueClient glueClient = GlueClient.builder()
-                .region(region)
-                .build();
-
-        getGlueTable(glueClient, dbName, tableName);
-        glueClient.close();
-    }
-
-    public static void getGlueTable(GlueClient glueClient, String dbName, String tableName) {
-        try {
-            GetTableRequest tableRequest = GetTableRequest.builder()
-                    .databaseName(dbName)
-                    .name(tableName)
-                    .build();
-
-            GetTableResponse tableResponse = glueClient.getTable(tableRequest);
-            Instant createDate = tableResponse.table().createTime();
-
-            // Convert the Instant to readable date.
-            DateTimeFormatter formatter = DateTimeFormatter.ofLocalizedDateTime(FormatStyle.SHORT)
-                    .withLocale(Locale.US)
-                    .withZone(ZoneId.systemDefault());
-
-            formatter.format(createDate);
-            System.out.println("The create date of the table is " + createDate);
-
-        } catch (GlueException e) {
-            System.err.println(e.awsErrorDetails().errorMessage());
-            System.exit(1);
-        }
-    }
-}
-// snippet-end:[glue.java2.get_table.main]
