+++ javav2/example_code/glue/src/main/java/com/example/glue/scenario/GlueScenario.java
@@ -1,0 +1,679 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.example.glue.scenario;
+
+// snippet-start:[glue.java2.scenario.import]
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.glue.GlueClient;
+import java.time.Instant;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+import java.time.format.FormatStyle;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Locale;
+import java.util.Map;
+import java.util.Scanner;
+import java.util.concurrent.TimeUnit;
+import software.amazon.awssdk.services.glue.model.DatabaseInput;
+import software.amazon.awssdk.services.glue.model.CreateDatabaseRequest;
+import software.amazon.awssdk.services.glue.model.GlueException;
+import software.amazon.awssdk.services.glue.model.GetCrawlerResponse;
+import software.amazon.awssdk.services.glue.model.GetCrawlerRequest;
+import software.amazon.awssdk.services.glue.model.CreateCrawlerRequest;
+import software.amazon.awssdk.services.glue.model.GetTablesResponse;
+import software.amazon.awssdk.services.glue.model.GetTablesRequest;
+import software.amazon.awssdk.services.glue.model.StartCrawlerRequest;
+import software.amazon.awssdk.services.glue.model.GetDatabaseRequest;
+import software.amazon.awssdk.services.glue.model.CrawlerTargets;
+import software.amazon.awssdk.services.glue.model.GetDatabaseResponse;
+import software.amazon.awssdk.services.glue.model.Table;
+import software.amazon.awssdk.services.glue.model.StartJobRunResponse;
+import software.amazon.awssdk.services.glue.model.StartJobRunRequest;
+import software.amazon.awssdk.services.glue.model.GetJobsRequest;
+import software.amazon.awssdk.services.glue.model.CreateJobRequest;
+import software.amazon.awssdk.services.glue.model.GetJobRunsResponse;
+import software.amazon.awssdk.services.glue.model.GetJobsResponse;
+import software.amazon.awssdk.services.glue.model.Job;
+import software.amazon.awssdk.services.glue.model.WorkerType;
+import software.amazon.awssdk.services.glue.model.JobCommand;
+import software.amazon.awssdk.services.glue.model.GetJobRunsRequest;
+import software.amazon.awssdk.services.glue.model.JobRun;
+import software.amazon.awssdk.services.glue.model.S3Target;
+import software.amazon.awssdk.services.glue.model.DeleteJobRequest;
+import software.amazon.awssdk.services.glue.model.DeleteDatabaseRequest;
+import software.amazon.awssdk.services.glue.model.DeleteCrawlerRequest;
+// snippet-end:[glue.java2.scenario.import]
+
+// snippet-start:[glue.java2.scenario.main]
+/**
+ *
+ * Before running this Java V2 code example, set up your development
+ * environment, including your credentials.
+ *
+ * For more information, see the following documentation topic:
+ *
+ * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
+ *
+ * To set up the resources, see this documentation topic:
+ *
+ * https://docs.aws.amazon.com/glue/latest/ug/tutorial-add-crawler.html
+ *
+ * This example performs the following tasks:
+ *
+ * 1. Create a database.
+ * 2. Create a crawler.
+ * 3. Get a crawler.
+ * 4. Start a crawler.
+ * 5. Get a database.
+ * 6. Get tables.
+ * 7. Create a job.
+ * 8. Start a job run.
+ * 9. List all jobs.
+ * 10. Get job runs.
+ * 11. Delete a job.
+ * 12. Delete a database.
+ * 13. Delete a crawler.
+ */
+
+public class GlueScenario {
+    public static final String DASHES = new String(new char[80]).replace("\0", "-");
+
+    public static void main(String[] args) throws InterruptedException {
+        final String usage = """
+
+                Usage:
+                    <iam> <s3Path> <cron> <dbName> <crawlerName> <jobName>\s
+
+                Where:
+                    iam - The ARN of the IAM role that has AWS Glue and S3 permissions.\s
+                    s3Path - The Amazon Simple Storage Service (Amazon S3) target that contains data (for example, CSV data).
+                    cron - A cron expression used to specify the schedule  (i.e., cron(15 12 * * ? *).
+                    dbName - The database name.\s
+                    crawlerName - The name of the crawler.\s
+                    jobName - The name you assign to this job definition.
+                    scriptLocation - The Amazon S3 path to a script that runs a job.
+                    locationUri - The location of the database
+                    bucketNameSc - The Amazon S3 bucket name used when creating a job
+                    """;
+
+        //if (args.length != 9) {
+        //    System.out.println(usage);
+        //    System.exit(1);
+       // }
+        Scanner scanner = new Scanner(System.in);
+        String iam = "arn:aws:iam::814548047983:role/AWSGlueServiceRoleDefault"  ; //args[0];
+        String s3Path = "s3://glue-demo-scott/read" ; //args[1];
+        String cron = "cron(15 12 * * ? *)" ; //args[2];
+        String dbName = "glue-db2" ; //args[3];
+        String crawlerName = "crawl2019" ; //args[4];
+        String jobName = "glue-job61" ; //args[5];
+        String scriptLocation = "s3://aws-glue-scripts-814548047983-us-east-1/PowerUserScott/flight_etl_job_script.py" ; //args[6];
+        String locationUri = "s3://crawler-public-us-east-1/flight/2016/csv/" ; //args[7];
+        String bucketNameSc = "glue-demo-scott" ; //args[8];
+
+        Region region = Region.US_EAST_1;
+        GlueClient glueClient = GlueClient.builder()
+                .region(region)
+                .build();
+        System.out.println(DASHES);
+        System.out.println("Welcome to the AWS Glue scenario.");
+        System.out.println("""
+            AWS Glue is a fully managed extract, transform, and load (ETL) service provided by Amazon 
+            Web Services (AWS). It is designed to simplify the process of building, running, and maintaining 
+            ETL pipelines, which are essential for data integration and data warehousing tasks.
+                        
+            One of the key features of AWS Glue is its ability to automatically discover and catalog data 
+            stored in various sources, such as Amazon S3, Amazon RDS, Amazon Redshift, and other databases. 
+            This cataloging process creates a central metadata repository, known as the AWS Glue Data Catalog, 
+            which provides a unified view of an organization's data assets. This metadata can then be used to 
+            create ETL jobs, which can be scheduled and run on-demand or on a regular basis.
+            
+            Lets get started.          
+                         
+            """);
+        waitForInputToContinue(scanner);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("1. Create a database.");
+        createDatabase(glueClient, dbName, locationUri);
+        waitForInputToContinue(scanner);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("2. Create a crawler.");
+        createGlueCrawler(glueClient, iam, s3Path, cron, dbName, crawlerName);
+        waitForInputToContinue(scanner);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("3. Get a crawler.");
+        getSpecificCrawler(glueClient, crawlerName);
+        waitForInputToContinue(scanner);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("4. Start a crawler.");
+        startSpecificCrawler(glueClient, crawlerName);
+        waitForInputToContinue(scanner);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("5. Get a database.");
+        getSpecificDatabase(glueClient, dbName);
+        waitForInputToContinue(scanner);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("*** Wait 5 min for the tables to become available");
+        TimeUnit.MINUTES.sleep(5);
+        System.out.println("6. Get tables.");
+        String myTableName = getGlueTables(glueClient, dbName);
+        waitForInputToContinue(scanner);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("7. Create a job.");
+        createJob(glueClient, jobName, iam, scriptLocation);
+        waitForInputToContinue(scanner);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("8. Start a Job run.");
+        startJob(glueClient, jobName, dbName, myTableName, bucketNameSc);
+        waitForInputToContinue(scanner);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("9. List all jobs.");
+        getAllJobs(glueClient);
+        waitForInputToContinue(scanner);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("10. Get job runs.");
+        getJobRuns(glueClient, jobName);
+        waitForInputToContinue(scanner);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("11. Delete a job.");
+        deleteJob(glueClient, jobName);
+        System.out.println("*** Wait 5 MIN for the " + crawlerName + " to stop");
+        TimeUnit.MINUTES.sleep(5);
+        waitForInputToContinue(scanner);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("12. Delete a database.");
+        deleteDatabase(glueClient, dbName);
+        waitForInputToContinue(scanner);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("Delete a crawler.");
+        deleteSpecificCrawler(glueClient, crawlerName);
+        waitForInputToContinue(scanner);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("Successfully completed the AWS Glue Scenario");
+        System.out.println(DASHES);
+    }
+
+    // snippet-start:[glue.java2.create_database.main]
+    /**
+     * Creates a Glue database with the specified name and location URI.
+     *
+     * @param glueClient The Glue client to use for the database creation.
+     * @param dbName     The name of the database to create.
+     * @param locationUri The location URI for the database.
+     */
+    public static void createDatabase(GlueClient glueClient, String dbName, String locationUri) {
+        try {
+            DatabaseInput input = DatabaseInput.builder()
+                    .description("Built with the AWS SDK for Java V2")
+                    .name(dbName)
+                    .locationUri(locationUri)
+                    .build();
+
+            CreateDatabaseRequest request = CreateDatabaseRequest.builder()
+                    .databaseInput(input)
+                    .build();
+
+            glueClient.createDatabase(request);
+            System.out.println(dbName + " was successfully created");
+
+        } catch (GlueException e) {
+            if (e.awsErrorDetails().errorMessage().equals("Database already exists.")) {
+                System.out.println("Database " + dbName + " already exists. Skipping creation.");
+            } else {
+                System.err.println(e.awsErrorDetails().errorMessage());
+                System.exit(1);
+            }
+        }
+    }
+    // snippet-end:[glue.java2.create_database.main]
+
+    // snippet-start:[glue.java2.create_crawler.main]
+    /**
+     * Creates a new AWS Glue crawler using the AWS Glue Java API.
+     *
+     * @param glueClient   the AWS Glue client used to interact with the AWS Glue service
+     * @param iam          the IAM role that the crawler will use to access the data source
+     * @param s3Path       the S3 path that the crawler will scan for data
+     * @param cron         the cron expression that defines the crawler's schedule
+     * @param dbName       the name of the AWS Glue database where the crawler will store the metadata
+     * @param crawlerName  the name of the crawler to be created
+     */
+    public static void createGlueCrawler(GlueClient glueClient,
+            String iam,
+            String s3Path,
+            String cron,
+            String dbName,
+            String crawlerName) {
+
+        try {
+            S3Target s3Target = S3Target.builder()
+                    .path(s3Path)
+                    .build();
+
+            List<S3Target> targetList = new ArrayList<>();
+            targetList.add(s3Target);
+            CrawlerTargets targets = CrawlerTargets.builder()
+                    .s3Targets(targetList)
+                    .build();
+
+            CreateCrawlerRequest crawlerRequest = CreateCrawlerRequest.builder()
+                    .databaseName(dbName)
+                    .name(crawlerName)
+                    .description("Created by the AWS Glue Java API")
+                    .targets(targets)
+                    .role(iam)
+                    .schedule(cron)
+                    .build();
+
+            glueClient.createCrawler(crawlerRequest);
+            System.out.println(crawlerName + " was successfully created");
+
+        } catch (GlueException e) {
+            if (e.awsErrorDetails().errorMessage().contains("already exists")) {
+                System.out.println("Crawler " + crawlerName + " already exists. Skipping creation.");
+            } else {
+                System.err.println(e.awsErrorDetails().errorMessage());
+                System.exit(1);
+            }
+        }
+    }
+    // snippet-end:[glue.java2.create_crawler.main]
+
+    // snippet-start:[glue.java2.get_crawler.main]
+    /**
+     * Retrieves a specific crawler from the AWS Glue service and waits for it to be in the "READY" state.
+     *
+     * @param glueClient  the AWS Glue client used to interact with the Glue service
+     * @param crawlerName the name of the crawler to be retrieved
+     */
+    public static void getSpecificCrawler(GlueClient glueClient, String crawlerName) {
+        try {
+            GetCrawlerRequest crawlerRequest = GetCrawlerRequest.builder()
+                    .name(crawlerName)
+                    .build();
+
+            boolean ready = false;
+            while (!ready) {
+                GetCrawlerResponse response = glueClient.getCrawler(crawlerRequest);
+                String status = response.crawler().stateAsString();
+                if (status.compareTo("READY") == 0) {
+                    ready = true;
+                }
+                Thread.sleep(3000);
+            }
+
+            System.out.println("The crawler is now ready");
+
+        } catch (GlueException | InterruptedException e) {
+            System.err.println(e.getMessage());
+            System.exit(1);
+        }
+    }
+    // snippet-end:[glue.java2.get_crawler.main]
+
+    // snippet-start:[glue.java2.start_crawler.main]
+    /**
+     * Starts a specific AWS Glue crawler.
+     *
+     * @param glueClient   the AWS Glue client to use for the crawler operation
+     * @param crawlerName  the name of the crawler to start
+     *
+     * @throws GlueException if there is an error starting the crawler
+     */
+    public static void startSpecificCrawler(GlueClient glueClient, String crawlerName) {
+        try {
+            StartCrawlerRequest crawlerRequest = StartCrawlerRequest.builder()
+                    .name(crawlerName)
+                    .build();
+
+            glueClient.startCrawler(crawlerRequest);
+            System.out.println(crawlerName + " was successfully started!");
+
+        } catch (GlueException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+    // snippet-end:[glue.java2.start_crawler.main]
+
+    // snippet-start:[glue.java2.get_database.main]
+    /**
+     * Retrieves the specific database from the AWS Glue service.
+     *
+     * @param glueClient     an instance of the AWS Glue client used to interact with the service
+     * @param databaseName   the name of the database to retrieve
+     * @throws GlueException if there is an error retrieving the database from the AWS Glue service
+     */
+    public static void getSpecificDatabase(GlueClient glueClient, String databaseName) {
+        try {
+            GetDatabaseRequest databasesRequest = GetDatabaseRequest.builder()
+                    .name(databaseName)
+                    .build();
+
+            GetDatabaseResponse response = glueClient.getDatabase(databasesRequest);
+            Instant createDate = response.database().createTime();
+
+            // Convert the Instant to readable date.
+            DateTimeFormatter formatter = DateTimeFormatter.ofLocalizedDateTime(FormatStyle.SHORT)
+                    .withLocale(Locale.US)
+                    .withZone(ZoneId.systemDefault());
+
+            formatter.format(createDate);
+            System.out.println("The create date of the database is " + createDate);
+
+        } catch (GlueException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+    // snippet-end:[glue.java2.get_database.main]
+
+    // snippet-start:[glue.java2.get_tables.main]
+    /**
+     * Retrieves the names of the tables in the specified Glue database.
+     *
+     * @param glueClient the Glue client to use for the operation
+     * @param dbName     the name of the Glue database to retrieve the table names from
+     * @return the name of the first table retrieved, or an empty string if no tables were found
+     */
+    public static String getGlueTables(GlueClient glueClient, String dbName) {
+        String myTableName = "";
+        try {
+            GetTablesRequest tableRequest = GetTablesRequest.builder()
+                    .databaseName(dbName)
+                    .build();
+
+            GetTablesResponse response = glueClient.getTables(tableRequest);
+            List<Table> tables = response.tableList();
+            if (tables.isEmpty()) {
+                System.out.println("No tables were returned");
+            } else {
+                for (Table table : tables) {
+                    myTableName = table.name();
+                    System.out.println("Table name is: " + myTableName);
+                }
+            }
+
+        } catch (GlueException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+        return myTableName;
+    }
+    // snippet-start:[glue.java2.get_tables.main]
+
+    // snippet-start:[glue.java2.start.job.main]
+    /**
+     * Starts a job run in AWS Glue.
+     *
+     * @param glueClient     the AWS Glue client to use for the job run
+     * @param jobName        the name of the Glue job to run
+     * @param inputDatabase  the name of the input database
+     * @param inputTable     the name of the input table
+     * @param outBucket      the URL of the output S3 bucket
+     *
+     * @throws GlueException if there is an error starting the job run
+     */
+    public static void startJob(GlueClient glueClient, String jobName, String inputDatabase, String inputTable,
+            String outBucket) {
+        try {
+            Map<String, String> myMap = new HashMap<>();
+            myMap.put("--input_database", inputDatabase);
+            myMap.put("--input_table", inputTable);
+            myMap.put("--output_bucket_url", outBucket);
+
+            StartJobRunRequest runRequest = StartJobRunRequest.builder()
+                    .workerType(WorkerType.G_1_X)
+                    .numberOfWorkers(10)
+                    .arguments(myMap)
+                    .jobName(jobName)
+                    .build();
+
+            StartJobRunResponse response = glueClient.startJobRun(runRequest);
+            System.out.println("The request Id of the job is " + response.responseMetadata().requestId());
+
+        } catch (GlueException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+    // snippet-end:[glue.java2.start.job.main]
+
+    // snippet-start:[glue.java2.create_job.main]
+    /**
+     * Creates a new AWS Glue job.
+     *
+     * @param glueClient      the AWS Glue client to use for the operation
+     * @param jobName         the name of the job to create
+     * @param iam             the IAM role to associate with the job
+     * @param scriptLocation  the location of the script to be used by the job
+     * @throws GlueException  if there is an error creating the job
+     */
+    public static void createJob(GlueClient glueClient, String jobName, String iam, String scriptLocation) {
+        try {
+            JobCommand command = JobCommand.builder()
+                    .pythonVersion("3")
+                    .name("glueetl")
+                    .scriptLocation(scriptLocation)
+                    .build();
+
+            CreateJobRequest jobRequest = CreateJobRequest.builder()
+                    .description("A Job created by using the AWS SDK for Java V2")
+                    .glueVersion("2.0")
+                    .workerType(WorkerType.G_1_X)
+                    .numberOfWorkers(10)
+                    .name(jobName)
+                    .role(iam)
+                    .command(command)
+                    .build();
+
+            glueClient.createJob(jobRequest);
+            System.out.println(jobName + " was successfully created.");
+
+        } catch (GlueException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+    // snippet-end:[glue.java2.create_job.main]
+
+    // snippet-start:[glue.java2.get_jobs.main]
+    /**
+     * Retrieves and prints information about all the jobs in the Glue data catalog.
+     *
+     * @param glueClient the Glue client used to interact with the AWS Glue service
+     */
+    public static void getAllJobs(GlueClient glueClient) {
+        try {
+            GetJobsRequest jobsRequest = GetJobsRequest.builder()
+                    .maxResults(10)
+                    .build();
+
+            GetJobsResponse jobsResponse = glueClient.getJobs(jobsRequest);
+            List<Job> jobs = jobsResponse.jobs();
+            for (Job job : jobs) {
+                System.out.println("Job name is : " + job.name());
+                System.out.println("The job worker type is : " + job.workerType().name());
+            }
+
+        } catch (GlueException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+    // snippet-end:[glue.java2.get_jobs.main]
+
+    // snippet-start:[glue.java2.get_job.main]
+    /**
+     * Retrieves the job runs for a given Glue job and prints the status of the job runs.
+     *
+     * @param glueClient the Glue client used to make API calls
+     * @param jobName    the name of the Glue job to retrieve the job runs for
+     */
+    public static void getJobRuns(GlueClient glueClient, String jobName) {
+        try {
+            GetJobRunsRequest runsRequest = GetJobRunsRequest.builder()
+                    .jobName(jobName)
+                    .maxResults(20)
+                    .build();
+
+            boolean jobDone = false;
+            while (!jobDone) {
+                GetJobRunsResponse response = glueClient.getJobRuns(runsRequest);
+                List<JobRun> jobRuns = response.jobRuns();
+                for (JobRun jobRun : jobRuns) {
+                    String jobState = jobRun.jobRunState().name();
+                    if (jobState.compareTo("SUCCEEDED") == 0) {
+                        System.out.println(jobName + " has succeeded");
+                        jobDone = true;
+
+                    } else if (jobState.compareTo("STOPPED") == 0) {
+                        System.out.println("Job run has stopped");
+                        jobDone = true;
+
+                    } else if (jobState.compareTo("FAILED") == 0) {
+                        System.out.println("Job run has failed");
+                        jobDone = true;
+
+                    } else if (jobState.compareTo("TIMEOUT") == 0) {
+                        System.out.println("Job run has timed out");
+                        jobDone = true;
+
+                    } else {
+                        System.out.println("*** Job run state is " + jobRun.jobRunState().name());
+                        System.out.println("Job run Id is " + jobRun.id());
+                        System.out.println("The Glue version is " + jobRun.glueVersion());
+                    }
+                    TimeUnit.SECONDS.sleep(5);
+                }
+            }
+
+        } catch (GlueException | InterruptedException e) {
+            System.err.println(e.getMessage());
+            System.exit(1);
+        }
+    }
+    // snippet-end:[glue.java2.get_job.main]
+
+    // snippet-start:[glue.java2.delete_job.main]
+    /**
+     * Deletes a Glue job.
+     *
+     * @param glueClient the Glue client to use for the operation
+     * @param jobName the name of the job to be deleted
+     * @throws GlueException if there is an error deleting the job
+     */
+    public static void deleteJob(GlueClient glueClient, String jobName) {
+        try {
+            DeleteJobRequest jobRequest = DeleteJobRequest.builder()
+                    .jobName(jobName)
+                    .build();
+
+            glueClient.deleteJob(jobRequest);
+            System.out.println(jobName + " was successfully deleted");
+
+        } catch (GlueException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+    // snippet-end:[glue.java2.delete_job.main]
+
+    // snippet-start:[glue.java2.delete_database.main]
+    /**
+     * Deletes a AWS Glue Database.
+     *
+     * @param glueClient    An instance of the AWS Glue client used to interact with the AWS Glue service.
+     * @param databaseName  The name of the database to be deleted.
+     *
+     * @throws GlueException If an error occurs while deleting the database.
+     */
+    public static void deleteDatabase(GlueClient glueClient, String databaseName) {
+        try {
+            DeleteDatabaseRequest request = DeleteDatabaseRequest.builder()
+                    .name(databaseName)
+                    .build();
+
+            glueClient.deleteDatabase(request);
+            System.out.println(databaseName + " was successfully deleted");
+
+        } catch (GlueException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+    // snippet-end:[glue.java2.delete_database.main]
+
+    // snippet-start:[glue.java2.delete_crawler.main]
+    /**
+     * Deletes a specific AWS Glue crawler.
+     *
+     * @param glueClient   the AWS Glue client object
+     * @param crawlerName  the name of the crawler to be deleted
+     * @throws GlueException if an error occurs during the deletion process
+     */
+    public static void deleteSpecificCrawler(GlueClient glueClient, String crawlerName) {
+        try {
+            DeleteCrawlerRequest deleteCrawlerRequest = DeleteCrawlerRequest.builder()
+                    .name(crawlerName)
+                    .build();
+
+            glueClient.deleteCrawler(deleteCrawlerRequest);
+            System.out.println(crawlerName + " was deleted");
+
+        } catch (GlueException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+    // snippet-end:[glue.java2.delete_crawler.main]
+
+    private static void waitForInputToContinue(Scanner scanner) {
+        while (true) {
+            System.out.println("");
+            System.out.println("Enter 'c' followed by <ENTER> to continue:");
+            String input = scanner.nextLine();
+
+            if (input.trim().equalsIgnoreCase("c")) {
+                System.out.println("Continuing with the program...");
+                System.out.println("");
+                break;
+            } else {
+                // Handle invalid input.
+                System.out.println("Invalid input. Please try again.");
+            }
+        }
+    }
+}
+// snippet-end:[glue.java2.scenario.main]
